requirements.txt
absl-py==0.8.1
astor==0.8.0
autopep8==1.4.4
cffi==1.13.1
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MySQLdb==0.2.0
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.24.1
gunicorn==19.9.0
h5py==2.10.0
imageio==2.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mysql-connector-python-rf==2.2.2
mysqlclient==1.4.4
networkx==2.3
numpy==1.17.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.0
protobuf==3.10.0
pycodestyle==2.5.0
pycparser==2.19
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
virtualenv==16.7.6
Werkzeug==0.16.0
wrapt==1.11.2
